Ajout du login et des droits sur les pages. Implémentation du token service qui permet d'avoir l'objet utilisateur connecté ainsi que le token

+++ src/app/assignment-detail/assignment-detail.component.ts
@@ -1,8 +1,9 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, Input, OnInit} from '@angular/core';
 import {Assignment} from "../assignments/model/assignment.model";
 import {AssignmentsService} from "../shared/assignments.service";
 import {ActivatedRoute, Router} from "@angular/router";
-import {AuthService} from "../shared/auth.service";
+import {MatSnackBar} from "@angular/material/snack-bar";
+import {TokenStorageService} from "../shared/token.service";
 
 @Component({
   selector: 'app-assignment-detail',
@@ -13,18 +14,26 @@ export class AssignmentDetailComponent implements OnInit {
   @Input() assignmentTransmis: Assignment | null | undefined;
   panelOpenState: boolean = false;
   isChecked: boolean = false;
+  private isLoggedIn: boolean;
+  private isAdmin: boolean;
+  private userName: string;
 
   constructor(private assignmentsService: AssignmentsService,
               private route: ActivatedRoute,
-              private authService: AuthService,
-              private router: Router) { }
+              private router: Router,
+              private _snackBar: MatSnackBar,
+              private tokenService: TokenStorageService) { }
 
   ngOnInit(): void {
-    this.getAssignment()
-  }
+    this.isLoggedIn = !!this.tokenService.getToken();
 
-  onSubmit(){
+    if(this.isLoggedIn){
+      const user = this.tokenService.getUser();
+      this.isAdmin = user.isAdmin;
 
+      this.userName = user.name;
+    }
+    this.getAssignment();
   }
 
   getAssignment() {
@@ -39,25 +48,36 @@ export class AssignmentDetailComponent implements OnInit {
   }
 
   handleUpdateAssignment() {
-    let assignment = this.assignmentTransmis ?? new Assignment();
-    assignment.rendu = !this.assignmentTransmis?.rendu;
-    this.assignmentsService.updateAssignment(assignment).subscribe(message => console.log(message));
-    this.router.navigate(["/home"]).then(r => console.log(r));
+    if(this.isLoggedIn){
+      if(new Date(this.assignmentTransmis?.dateDeRendu ?? new Date()) > new Date()) {
+        let assignment = this.assignmentTransmis ?? new Assignment();
+        assignment.rendu = !this.assignmentTransmis?.rendu;
+        this.assignmentsService.updateAssignment(assignment).subscribe(message => console.log(message));
+        this.router.navigate(["/home"]).then(r => console.log(r));
+      } else {
+        this.isChecked = false;
+        this.openSnackBar("La date de rendu est passée, vous ne pouvez pas modifier le devoir");
+      }
+    } else{
+      this.router.navigate(["/connexion"]);
+    }
   }
 
   handleDeleteAssignment() {
-    this.assignmentsService.deleteAssignment(this.assignmentTransmis ?? new Assignment()).subscribe((message) => {
-      console.log(message);
-      this.router.navigate(["/home"]).then(r => console.log(r));
+    if(this.isLoggedIn) {
+      this.assignmentsService.deleteAssignment(this.assignmentTransmis ?? new Assignment()).subscribe((message) => {
+        console.log(message);
+        this.router.navigate(["/home"]).then(r => console.log(r));
+      });
+    } {
+      this.router.navigate(["/connexion"]);
+    }
+  }
+
+  openSnackBar(message: string) {
+    this._snackBar.open(message, 'Terminé', {
+      horizontalPosition: "right",
+      verticalPosition: "top",
     });
-  }
-
-  onClickEdit(){
-    this.router.navigate(["/assignment", this.assignmentTransmis?._id, 'edit'],
-      {queryParams: {nom: this.assignmentTransmis?.nom}, fragment: 'edition'});
-  }
-
-  isAdmin(){
-    return this.authService.loggedIn;
   }
 }
